adding image viewer widget
a widget that can convert a np.array into a qimage for display using
graphics view, this is similar to how pyqtgraph is, but not one big
class. missing menu items, colourmap. not sure if it could be used for a
video, need to figure out performance.
the scene holds some information about layers, which may remove at some
point.
the ImageViewComponent is effectively the top layer, holding the scene
and viewer references and some convience methods.
its coming from another project